backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
psycopg2-binary
requests
passlib[bcrypt]
python-jose[cryptography]